decode afc directly to floats

+++ sms/afc.hh
@@ -4,5 +4,4 @@
 
 #include <vector>
 
-std::vector<int16_t> afc_decode(const void* data, size_t size,
-    bool small_frames);
+std::vector<float> afc_decode(const void* data, size_t size, bool small_frames);
